// src/types/films.interface.ts
export interface FilmInterface {
  uid: string
  filmImg: string
  slug: string
  name: string
  originalName: string
  ratingIMDb: string
  description: string
  type: string
  status: string
  year: string
  country: string
  genre: string
  director: string
  actors: string
  iframeUrl: string
}